application/requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-limiter==3.3.1
greenlet==2.0.2
importlib-metadata==6.6.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
libpcap==1.11.0b7
MarkupSafe==2.1.2
packaging==23.1
pkg-about==1.0.8
scapy==2.5.0
SQLAlchemy==2.0.12
tomli==2.0.1
typing_extensions==4.5.0
Werkzeug==2.3.3
WTForms==3.0.1
captcha==0.4
google-auth==2.18.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0